chore: Folder Grouping
MVVM 패턴적용을 위한 폴더 그룹핑 & 주석제거
