+++ iOS/SideDish/SideDish/Presentaion/DetailBanchan/VIew/DetailBanchanViewController.swift
@@ -1,0 +1,15 @@
+//
+//  DetailBanchanViewController.swift
+//  SideDish
+//
+//  Created by 심영민 on 2021/04/20.
+//
+
+import UIKit
+
+class DetailBanchanViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+    }
+}
