+++ iOS/SideDish/SideDish/SupportingFiles/Assets.xcassets/AppIcon.appiconset/Contents.json
