+++ iOS/SideDish/SideDish/SupportingFiles/SceneDelegate.swift
@@ -1,0 +1,19 @@
+//
+//  SceneDelegate.swift
+//  SideDish
+//
+//  Created by 심영민 on 2021/04/20.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let _ = (scene as? UIWindowScene) else { return }
+    }
+}
+
